Added ``Error`` and ``OnUseSkill`` handler
- Fixed injection
- Changed from f32 to f64

+++ src/widgets/av_damage_plot_widget.rs
@@ -21,7 +21,7 @@ impl DamageAnalyzer {
                         .iter()
                         .map(|dmg_map| dmg_map.get(name).unwrap())
                         .copied()
-                        .collect::<Vec<f32>>();
+                        .collect::<Vec<f64>>();
 
                     let points = data_buffer.av_history
                         .iter()
